No functions take indices 2

+++ src/cluster.rs
@@ -5,7 +5,7 @@ pub fn sweep_line_cluster(
     starts: &[i64],
     ends: &[i64],
     slack: i64,
-) -> (Vec<usize>, Vec<usize>) {
+) -> (Vec<i64>, Vec<usize>) {
     let mut indices = Vec::with_capacity(chrs.len());
     let mut cluster_ids = Vec::with_capacity(chrs.len());
 
@@ -16,7 +16,7 @@ pub fn sweep_line_cluster(
     let events = sorts::build_sorted_events_single_collection(chrs, starts, ends, slack);
 
     let mut current_chr: i64 = events.first().unwrap().chr;
-    let mut current_cluster: usize = 0;
+    let mut current_cluster: i64 = 0;
     let mut active_intervals: i64 = 0;
 
     for e in events {
